chat-server-python/requirements.txt
websockets==12.0
python-dotenv==1.0.0
openai==1.2.3
python-decouple==3.8
httpx>=0.24.1
